Move wrapper class out of S3AsyncClient_Instrumentation
